Check for subprocess errors and show exceptions from threads
Calling `subprocess.call()` ignores the return value from the
subprocess, so it wouldn't complain when Blender or Luminance HDR fails
for some reason.

Python doesn't report any exceptions that occur in a thread, so it's
important to check the result of a `Future` once it's done.

As a final touch, I replaced the use of the private `t._state ==
"FINISHED"` with the proper API call `t.done()`.